+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 ### Changed
+- Capture Error returns optional
 ### Removed
 
 ## [1.4.0] - 2023-03-12